+++ gostringsr2/__init__.py
@@ -6,4 +6,4 @@ __author__ = """Jonathan Wrightsell"""
 __email__ = "jonathan.wrightsell@carvesystems.com"
 __version__ = "1.0.0"
 
-from .gostringsr2 import GoStringsR2
+from .gostringsr2 import *