+++ gostringsr2/cli.py
@@ -1,6 +1,7 @@
 import sys
 import click
-from . import GoStringsR2
+from os import path
+from . import GoStringsR2, GoStringsR2Error
 
 
 def printe(*args, **kwargs):
@@ -9,41 +10,42 @@ def printe(*args, **kwargs):
 
 @click.command()
 @click.argument("file")
-def main(file):
+@click.option(
+    "-n", "length", is_flag=False, default=4, help="minimum length, default=4"
+)
+@click.option("-v", "verbose", is_flag=True, help="verbose")
+@click.option("-u", "utf8", is_flag=True, help="utf8 encoding instead of ascii")
+def main(file, length, verbose, utf8):
 
-    g = GoStringsR2(file)
+    if not path.isfile(file):
+        printe("invalid file {}".format(file))
+        return 1
 
-    g.load()
+    g = GoStringsR2(file, verbose)
 
-    printe("\nanalyzing: {}".format(g.file_info()))
+    try:
+        g.load()
 
-    strtab = g.get_string_table()
-    if strtab is None:
-        printe("error: couldn't find string table")
-        exit(1)
+        encoding = "ascii"
+        if utf8:
+            encoding = "utf8"
 
-    xrefs = g.get_cross_refs()
+        go_strings = g.get_strings(length, encoding)
+        # array of [address, decoded length, string]
 
-    if xrefs is not None:
-        str_refs = g.process_xrefs(xrefs, strtab["startaddr"], strtab["endaddr"])
+        for go_string in go_strings:
+            s_addr, s_len, s_val = go_string
+            if verbose:
+                print('0x{:x} : [{}] : {}'.format(s_addr, s_len, s_val))
+            else:
+                print(s_val)
 
-        if len(str_refs) == 0:
-            printe("error: no cross-references found to the string table")
-            results = None
-        else:
-            results = g.find_strings(str_refs, strtab["startaddr"], strtab["data"])
+        g.kill()
 
-        if results is not None:
-            r = len(results) - 1
-            while r >= 0:
-                print(results[r][2].decode("ascii", errors="ignore"))
-                r -= 1
-    else:
-        printe("error: no cross-references found; cannot locate strings.")
+        return 0
 
-    g.kill()
-
-    return 0
+    except GoStringsR2Error as err:
+        printe("gostringsr2 error: {}".format(err))
 
 
 if __name__ == "__main__":