+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## next
+
+- Use `p8` instead of `pr` to get raw data
+- Fix bug in is_a_string_ref()
+- Use quietr2pipe to kill r2's stderr during run()
+
 ## 1.0.0
 
 - Initial release